# backend/api/openAPI.yaml
openapi: 3.0.3
info:
  title: YoloFarm API - Đồ án Thực tập Đa ngành (CO3109)
  description: |
    Tài liệu đặc tả API cho hệ thống Nông nghiệp thông minh YoloFarm.
    
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Spring Boot Server
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Đăng ký tài khoản Nông dân
      security: [] 
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              KichBanChuan:
                summary: Đầy đủ thông tin hợp lệ
                value:
                  username: nongdan_01
                  password: SecurePassword123!
                  email: nongdan01@gmail.com
              KichBanThieuEmail:
                summary: Cố tình không nhập Email
                value:
                  username: nongdan_01
                  password: SecurePassword123!
      responses:
        '201':
          description: Đăng ký thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user_id:
                    type: string
                    format: uuid
              examples:
                ThanhCong:
                  summary: Trả về ID của user mới
                  value:
                    message: "Đăng ký tài khoản thành công!"
                    user_id: "123e4567-e89b-12d3-a456-426614174000"
        '400':
          description: Dữ liệu đầu vào không hợp lệ (Bad Request)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                LoiMatKhau:
                  summary: Lỗi mật khẩu quá yếu
                  value:
                    code: 400
                    message: Validation Failed
                    details: "Mật khẩu phải chứa ít nhất 8 ký tự, bao gồm chữ và số."
                LoiThieuTruong:
                  summary: Lỗi bỏ trống trường bắt buộc
                  value:
                    code: 400
                    message: Validation Failed
                    details: "Trường 'email' không được để trống."
        '409':
          description: Xung đột dữ liệu (Conflict)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                LoiTrungUsername:
                  summary: Tên đăng nhập đã tồn tại
                  value:
                    code: 409
                    message: Conflict
                    details: "Tên đăng nhập 'nongdan_01' đã được sử dụng."
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Đăng nhập hệ thống
      security: [] 
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              DangNhapChuan:
                summary: Nhập đúng tài khoản mật khẩu
                value:
                  username: nongdan_01
                  password: SecurePassword123!
      responses:
        '200':
          description: Đăng nhập thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
              examples:
                DangNhapThanhCong:
                  summary: Trả về JWT Token
                  value:
                    access_token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOi..."
                    token_type: "Bearer"
                    expires_in: 86400
                    role: "FARMER"
        '401':
          description: Sai tên đăng nhập hoặc mật khẩu (Unauthorized)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                SaiMatKhau:
                  summary: Nhập sai mật khẩu
                  value:
                    code: 401
                    message: Unauthorized
                    details: "Tài khoản hoặc mật khẩu không chính xác."
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Lấy thông tin cá nhân hiện tại
      responses:
        '200':
          description: Chi tiết Profile người dùng
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
              examples:
                ProfileNongDan:
                  summary: Profile của Nông Dân
                  value:
                    id: "123e4567-e89b-12d3-a456-426614174000"
                    username: "nongdan_01"
                    email: "nongdan01@gmail.com"
                    role: "FARMER"
                    created_at: "2023-10-01T12:00:00Z"
                ProfileAdmin:
                  summary: Profile của Admin
                  value:
                    id: "999e4567-e89b-12d3-a456-426614179999"
                    username: "admin_tong"
                    email: "admin@yolofarm.com"
                    role: "ADMIN"
                    created_at: "2023-01-01T08:00:00Z"
        '401':
          description: Token không hợp lệ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                TokenHetHan:
                  summary: Token đã hết hạn
                  value:
                    code: 401
                    message: Unauthorized
                    details: "JWT Token đã hết hạn. Vui lòng đăng nhập lại."

  # ==========================================
  # 2. FARMS (NÔNG TRẠI)
  # ==========================================
...
  /api/v1/devices/{deviceId}:
    patch:
      tags: [Devices]
      summary: (FARMER) Đổi tên hiển thị của Thiết bị
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
            examples:
              DoiTen:
                summary: Đổi tên thiết bị cho dễ nhớ
                value:
                  name: "Bơm tưới giàn Lan"
      responses:
        '200':
          description: Cập nhật tên thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
  /api/v1/devices/{deviceId}/remove-requests:
    post:
      tags: [Devices]
      summary: (FARMER) Gửi yêu cầu gỡ bỏ/trả lại thiết bị
      description: API này sẽ đổi status của thiết bị thành PENDING_REMOVAL.
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Gửi yêu cầu thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              examples:
                YeuCauXoa:
                  summary: Đã báo cho Admin
                  value:
                    message: "Đã gửi yêu cầu gỡ bỏ thiết bị. Vui lòng chờ Admin xác nhận thu hồi."
  /api/v1/admin/devices/{deviceId}:
    delete:
      tags: [Devices]
      summary: (ADMIN) Xóa hoàn toàn thiết bị & Thu hồi Feed Key
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Đã xóa thiết bị và giải phóng Feed Key
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              examples:
                XoaThanhCong:
                  summary: Hoàn tất thu hồi
                  value:
                    message: "Đã thu hồi Feed Key và xóa thiết bị khỏi hệ thống."
  /api/v1/devices/requests:
    post:
      tags: [Devices]
      summary: (FARMER) Yêu cầu thêm thiết bị mới
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceRequest'
            examples:
              YeuCauTuModel:
                summary: Chọn Model Máy Bơm từ Dropdown
                value:
                  farm_id: "888e4567-e89b-12d3-a456-426614175555"
                  model_id: "m11e4567-e89b-12d3-a456-426614171111"
                  name: "Máy bơm nước hồ 1"
      responses:
        '201':
          description: Đã gửi yêu cầu thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceResponse'
              examples:
                ThanhCong:
                  summary: Trạng thái PENDING
                  value:
                    id: "d11e4567-e89b-12d3-a456-426614171111"
                    farm_id: "888e4567-e89b-12d3-a456-426614175555"
                    model_id: "m11e4567-e89b-12d3-a456-426614171111"
                    name: "Máy bơm nước hồ 1"
                    status: "PENDING"
                    connection_status: "OFFLINE"
                    operating_mode: "MANUAL"
        '403':
          description: Không có quyền thao tác trên Nông trại này
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                LoiQuyen:
                  summary: Nông dân điền farm_id của người khác
                  value:
                    code: 403
                    message: Forbidden
                    details: "Bạn không có quyền thêm thiết bị vào nông trại này."
  /api/v1/admin/devices/requests:
    get:
      tags: [Devices]
      summary: (ADMIN) Xem danh sách thiết bị chờ duyệt
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
          examples:
            LocChoDuyet:
              summary: Lọc các đơn PENDING
              value: "PENDING"
      responses:
        '200':
          description: Danh sách đơn yêu cầu
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DeviceResponse'
              examples:
                DanhSachCho:
                  summary: Có 1 thiết bị đang chờ duyệt
                  value:
                    - id: "d11e4567-e89b-12d3-a456-426614171111"
                      farm_id: "888e4567-e89b-12d3-a456-426614175555"
                      name: "Máy bơm nước hồ 1"
                      device_type: "ACTUATOR"
                      status: "PENDING"
                      adafruit_feed_key: null
        '403':
          description: Chỉ Admin mới được gọi API này
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                KhongPhaiAdmin:
                  summary: Role FARMER cố tình gọi API Admin
                  value:
                    code: 403
                    message: Forbidden
                    details: "Access Denied. Cần quyền ADMIN."
...
  /api/v1/devices/{deviceId}/telemetry:
    get:
      tags: [Telemetry]
      summary: Lấy lịch sử dữ liệu của một Cảm biến
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: start_time
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: end_time
          in: query
          required: true
          schema:
            type: string
            format: date-time
        - name: aggregate
          in: query
          required: false
          schema:
            type: string
            enum: [15m, 1h, 1d]
          description: "Gom nhóm dữ liệu (VD: trung bình mỗi giờ) để tránh quá tải trình duyệt"
      responses:
        '200':
          description: Trả về mảng dữ liệu thời gian thực
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TelemetryDataPoint'
              examples:
                DuLieuNhietDo:
                  summary: Biểu đồ Nhiệt độ theo giờ
                  value:
                    - time: "2023-10-01T08:00:00Z"
                      value: 32.5
                    - time: "2023-10-01T09:00:00Z"
                      value: 34.1
                KhongCoDuLieu:
                  summary: Cảm biến bị mất kết nối
                  value: []
        '400':
          description: Lỗi tham số thời gian
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                LoiThoiGian:
                  summary: start_time lớn hơn end_time
                  value:
                    code: 400
                    message: Bad Request
                    details: "Thời gian bắt đầu (start_time) không thể lớn hơn thời gian kết thúc (end_time)."

  # ==========================================
  # 6. CONTROL (ĐIỀU KHIỂN THỦ CÔNG)
  # ==========================================
  /api/v1/devices/{deviceId}/command:
    post:
      tags: [Control]
      summary: Gửi lệnh điều khiển xuống Thiết bị thực thi (Actuator)
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                command:
                  type: string
                  enum: [ON, OFF]
            examples:
              LenhBat:
                summary: Bật máy bơm
                value:
                  command: "ON"
              LenhTat:
                summary: Tắt máy bơm
                value:
                  command: "OFF"
      responses:
        '200':
          description: Đã gửi lệnh xuống MQTT Broker
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              examples:
                ThanhCong:
                  summary: Phản hồi thành công
                  value:
                    message: "Lệnh [ON] đã được gửi tới thiết bị thành công."
        '400':
          description: Lỗi logic thiết bị
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                LoiSaiLoaiThietBi:
                  summary: Cố tình gửi lệnh cho Cảm biến
                  value:
                    code: 400
                    message: Bad Request
                    details: "Không thể gửi lệnh điều khiển. Thiết bị này là SENSOR, không phải ACTUATOR."
                LoiChuaActive:
                  summary: Thiết bị chưa được Admin duyệt
                  value:
                    code: 400
                    message: Bad Request
                    details: "Thiết bị đang ở trạng thái PENDING. Vui lòng đợi Admin cấp phát Feed Key."
                LoiDangOCheDoAuto:
                  summary: Đang ở chế độ Tự Động
                  value:
                    code: 400
                    message: Bad Request
                    details: "Thiết bị đang ở chế độ Tự Động (AUTO). Vui lòng chuyển sang Thủ Công (MANUAL) trước khi tự điều khiển."

  # ==========================================
  # 7. AI ANALYSIS (TRÍ TUỆ NHÂN TẠO)
  # ==========================================
...
  /api/v1/devices/{deviceId}/mode:
    patch:
      tags: [Control]
      summary: (FARMER) Đổi chế độ hoạt động (AUTO / MANUAL)
      parameters:
        - name: deviceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                operating_mode:
                  type: string
                  enum: [AUTO, MANUAL]
            examples:
              DoiSangThuCong:
                summary: Tắt tự động hóa để điều khiển tay
                value:
                  operating_mode: "MANUAL"
      responses:
        '200':
          description: Chuyển chế độ thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              examples:
                ThanhCong:
                  summary: Phản hồi
                  value:
                    message: "Đã chuyển thiết bị sang chế độ MANUAL."

# ==========================================
# CÁC LƯỢC ĐỒ (SCHEMAS) TÁI SỬ DỤNG
# ==========================================
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
        details:
          type: string
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        email:
          type: string
    LoginRequest:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
        expires_in:
          type: integer
        role:
          type: string
    UserProfile:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        email:
          type: string
        role:
          type: string
        created_at:
          type: string
...
    DeviceRequest:
      type: object
      required: [farm_id, model_id, name]
      properties:
        farm_id:
          type: string
          format: uuid
        model_id:
          type: string
          format: uuid
          description: "ID của khuôn mẫu thiết bị (Lấy từ Device Models)"
        name:
          type: string
          description: "Tên hiển thị do Nông dân đặt"
    DeviceResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        farm_id:
          type: string
          format: uuid
        model_id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [PENDING, ACTIVE, REJECTED]
        adafruit_feed_key:
          type: string
          nullable: true
        connection_status:
          type: string
          enum: [ONLINE, OFFLINE]
          description: "FR3.4: Trạng thái kết nối"
        last_seen:
          type: string
          format: date-time
          nullable: true
        operating_mode:
          type: string
          enum: [AUTO, MANUAL]
          description: "FR4: Chế độ điều khiển"
        is_active:
          type: boolean

    # --- SCHEMAS NHÓM TỰ ĐỘNG HÓA (RULES) ---
...
    TelemetryDataPoint:
      type: object
      properties:
        time:
          type: string
          format: date-time
          description: "Thời gian ghi nhận dữ liệu (từ InfluxDB/Hypertable)"
        value:
          type: number
          format: float
          description: "Giá trị đo được (Nhiệt độ, độ ẩm...)"

    # --- SCHEMAS NHÓM AI ANALYSIS ---
...