# backend/api/src/main/java/com/yoloFarm/api/dto/request/LoginRequest.java
package com.yoloFarm.api.dto.request;

import jakarta.validation.constraints.NotBlank;
import lombok.Data;

@Data
public class LoginRequest {
    @NotBlank(message = "Username cannot be empty")
    private String username;

    @NotBlank(message = "Password cannot be empty")
    private String password;
}